More changes as per review
